+++ TagsCloudContainer/CloudLayouters/ICloudLayouter.cs
@@ -1,6 +1,6 @@
 ﻿using System.Drawing;
 
-namespace TagsCloudContainer
+namespace TagsCloudContainer.CloudLayouters
 {
     interface ICloudLayouter
     {
