--- TagsCloudContainer/SimpleWordCounterTests.cs
@@ -1,15 +1,0 @@
-﻿using NUnit.Framework;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace TagsCloudContainer
-{
-    [TestFixture]
-    class SimpleWordCounterTests
-    {
-
-    }
-}
