+++ TagsCloudContainer/TagsCloudGenerator.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using TagsCloudContainer.CloudLayouters;
+using TagsCloudContainer.WordCounters;
 
 namespace TagsCloudContainer
 {
